+++ src/client/classes/message.class.js
@@ -18,7 +18,7 @@
  */
 
 import _ from 'lodash';
-const assert = require( 'assert' ).strict; // up to nodejs v16.x
+import { strict as assert } from 'node:assert'; // up to nodejs v16.x
 import mix from '@vestergaard-company/js-mixin';
 
 import { Logger } from 'meteor/pwix:logger';