+++ src/client/interfaces/ifield-run.iface.js
@@ -5,7 +5,7 @@
  */
 
 import _ from 'lodash';
-const assert = require( 'assert' ).strict;
+import { strict as assert } from 'node:assert';
 import { DeclareMixin } from '@vestergaard-company/js-mixin';
 
 import { Blaze } from 'meteor/blaze';