+++ src/client/interfaces/ifield-spec.iface.js
@@ -20,7 +20,7 @@
  */
 
 import _ from 'lodash';
-const assert = require( 'assert' ).strict;
+import { strict as assert } from 'node:assert';
 import { DeclareMixin } from '@vestergaard-company/js-mixin';
 
 import { Logger } from 'meteor/pwix:logger';